Rename geojson package
